:sparkles:[#1]react 적용
