+++ public/index.html
@@ -7,6 +7,6 @@
   <title>아워코리아트립</title>
 </head>
 <body>
-  
+  <div id="root"></div>
 </body>
 </html>
