+++ src/App.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const App = () => {
+  return (
+    <div>
+      hi
+    </div>
+  )
+}
+
+export default App;
