+++ src/index.js
@@ -1,38 +1,5 @@
-import _ from 'lodash';
-import './style.css';
-import Icon from './icon.png';
-import printMe from './print.js';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
 
-function component() {
-  const element = document.createElement('div');
-
-  // Lodash, currently included via a script, is required for this line to work
-  element.innerHTML = _.join(['Hello', 'webpack'], ' ');
-  element.classList.add('hello');
-
-  // Add the image to our existing div.
-  const myIcon = new Image();
-  myIcon.src = Icon;
-
-  element.appendChild(myIcon);
-
-  const btn = document.createElement('button');
-  btn.innerHTML = 'Click me and check the console!';
-  btn.onclick = printMe;
-
-  element.appendChild(btn);
-
-  return element;
-}
-
-let element = component();
-document.body.appendChild(element);
-
-if (module.hot) {
-  module.hot.accept('./print.js', function() {
-    console.log('Accepting the update printMe module!');
-    document.body.removeChild(element);
-    element = component();
-    document.body.appendChild(element);
-  })
-}
+ReactDOM.render(<App />, document.getElementById('root'));
